Moved route finder algorithm to its own service and tests refactored accordingly
