+++ src/main/java/mx/erickb/shipping/service/IRoutesManagerService.java
@@ -1,0 +1,10 @@
+package mx.erickb.shipping.service;
+
+import mx.erickb.shipping.util.RoutesGraph;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface IRoutesManagerService {
+    Optional<List<String>> findOptimalRoute(RoutesGraph routesGraph, String origin, String destination);
+}
