+++ src/main/java/mx/erickb/shipping/util/RouteUtils.java
@@ -4,12 +4,12 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import mx.erickb.shipping.exception.InvalidResponseException;
 import mx.erickb.shipping.model.Route;
-import org.apache.commons.lang3.tuple.Pair;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
 
-import java.util.*;
+import java.util.Arrays;
+import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
@@ -45,39 +45,5 @@ public class RouteUtils {
             routesGraph.addRoute(route.getFrom(), route.getTo(), route.getDistance());
         }
         return routesGraph;
-    }
-
-    // Breadth-first search using priority queue
-    public Optional<List<String>> findOptimalRoute(RoutesGraph routesGraph, String origin, String destination) {
-        // Priority queue that returns min instead of max element
-        PriorityQueue<Pair<List<String>, Integer>> search
-                = new PriorityQueue<>(Comparator.comparing(Pair::getRight));
-
-        // Inserting root node (BFS starting point)
-        search.add(Pair.of(Collections.singletonList(origin), 0));
-
-        while (!search.isEmpty()) {
-            Pair<List<String>, Integer> currentNode = search.poll();
-            List<String> currentRoute = currentNode.getLeft();
-            String currentCity = currentRoute.get(currentRoute.size() - 1);
-            if (currentCity.equals(destination)) {
-                return Optional.of(currentRoute);
-            }
-
-            // Add all the (non-visited) child nodes to the  search
-            Integer currentCost = currentNode.getRight();
-            for (String connectedCity : routesGraph.getConnectedCities(currentCity)) {
-                if (currentRoute.contains(connectedCity)) {
-                    // Skip if already visited in current route (avoid cycles)
-                    continue;
-                }
-                // Include connected city in current route
-                List<String> updatedRoute = new ArrayList<>(currentRoute);
-                updatedRoute.add(connectedCity);
-                Integer connectionCost = routesGraph.getDistance(currentCity, connectedCity);
-                search.add(Pair.of(updatedRoute, currentCost + connectionCost));
-            }
-        }
-        return Optional.empty();
     }
 }
